Reorganise Middleware service to be within the Http namespace
