+++ Polyel/src/Http/Middleware/CsrfTokenVerifier.class.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Polyel\Middleware;
+namespace Polyel\Http\Middleware;
 
 use Polyel\Session\Session;
