+++ app/Middleware/ValidateCsrfTokenMiddleware.class.php
@@ -2,7 +2,7 @@
 
 namespace App\Middleware;
 
-use Polyel\Middleware\CsrfTokenVerifier as PolyelCsrfTokenVerifierMiddleware;
+use Polyel\Http\Middleware\CsrfTokenVerifier as PolyelCsrfTokenVerifierMiddleware;
 
 class ValidateCsrfTokenMiddleware extends PolyelCsrfTokenVerifierMiddleware
 {
